+++ .idea/dataSources/8d71cc62-5966-4caf-b17f-09eaa6b7dcd7/storage_v2/_src_/schema/mysql.osA4Bg.meta
@@ -1,0 +1,2 @@
+#n:mysql
+!<md> [null, 0, null, null, -2147483648, -2147483648]
