--- src/main/java/com/emiliodevesa/kot/HibernateUtil.java
@@ -1,75 +1,0 @@
-package kot;
-
-import org.hibernate.Session;
-import org.hibernate.SessionFactory;
-import org.hibernate.boot.registry.StandardServiceRegistryBuilder;
-import org.hibernate.cfg.Configuration;
-import org.hibernate.cfg.Environment;
-import org.hibernate.service.ServiceRegistry;
-
-import java.util.Properties;
-
-/**
- * Una clase que permite la creación de sesiones de conexión a una BBDD a través de Hibernate ORM
- * @author Emilio Devesa
- * @version 0.1
- */
-public class HibernateUtil {
-
-    private static SessionFactory sessionFactory;
-    private static Session session;
-
-    /**
-     * Crea la factoria de sesiones
-     */
-    public static void buildSessionFactory() {
-        // Hibernate settings equivalent to hibernate.cfg.xml's properties
-        Properties settings = new Properties();
-        settings.put(Environment.DRIVER, "com.mysql.cj.jdbc.Driver");
-        settings.put(Environment.URL, "jdbc:mysql://localhost/KOTDB");
-        settings.put(Environment.USER, "root");
-        settings.put(Environment.PASS, "adminpwd");
-        settings.put(Environment.DIALECT, "org.hibernate.dialect.MySQL5Dialect");
-        settings.put(Environment.SHOW_SQL, "true");
-        settings.put(Environment.CURRENT_SESSION_CONTEXT_CLASS, "thread");
-        //settings.put(Environment.HBM2DDL_AUTO, "create-drop");
-        Configuration configuration = new Configuration();
-        configuration.setProperties(settings);
-        configuration.configure();
-        // Se registran las clases que hay que mapear con cada tabla de la base de datos
-        configuration.addAnnotatedClass(Issue.class);
-        ServiceRegistry serviceRegistry = new StandardServiceRegistryBuilder().applySettings(
-                configuration.getProperties()).build();
-        sessionFactory = configuration.buildSessionFactory(serviceRegistry);
-    }
-
-    /**
-     * Abre una nueva sesión
-     */
-    public static void openSession() {
-        session = sessionFactory.openSession();
-    }
-
-    /**
-     * Devuelve la sesión actual
-     * @return La sesión de conexión a la base de datos, a través de la cual hacer consultas
-     */
-    public static Session getCurrentSession() {
-        if ((session == null) || (!session.isOpen()))
-            openSession();
-        return session;
-    }
-
-    /**
-     * Cierra Hibernate
-     */
-    public static void closeSessionFactory() {
-        if (session != null)
-            session.close();
-        if (sessionFactory != null)
-            sessionFactory.close();
-    }
-
-
-}
-
