+++ src/main/java/com/emiliodevesa/kot/MainController.java
@@ -1,0 +1,14 @@
+package com.emiliodevesa.kot;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class MainController {
+
+    @GetMapping("")
+    public String showMainPage() {
+        return "index";
+    }
+
+}
