+++ src/main/java/com/emiliodevesa/kot/core/IssueRepository.java
@@ -1,0 +1,7 @@
+package com.emiliodevesa.kot.core;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface IssueRepository extends CrudRepository<Issue, Long> {
+
+}
