--- src/main/java/com/emiliodevesa/kot/hibernate.cfg.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE hibernate-configuration PUBLIC "-//Hibernate/Hibernate Configuration DTD 3.0//EN"
-        "http://www.hibernate.org/dtd/hibernate-configuration.dtd">
-<hibernate-configuration>
-    <session-factory name="session-factory">
-    </session-factory>
-</hibernate-configuration>
