+++ src/main/resources/templates/index.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <title>KOT: Keep On Track</title>
+    <link rel="stylesheet" type="text/css" th:href="@{/webjars/bootstrap/css/bootstrap.min.css}" />
+</head>
+<body>
+<div class="container-fluid text-center">
+    <h1>KOT: Keep On Track</h1>
+    <a class="m-2" th:href="@{/issue/new}">New Issue</a>
+    <a class="m-2" th:href="@{/issue/list}">Issue List</a>
+</div>
+</body>
+</html>
