+++ src/main/resources/templates/newIssue.html
@@ -1,0 +1,47 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org" xmlns="http://www.w3.org/1999/html">
+<head>
+    <meta charset="UTF-8">
+    <title>KOT: New Issue</title>
+    <link rel="stylesheet" type="text/css" th:href="@{/webjars/bootstrap/css/bootstrap.min.css}" />
+</head>
+<body>
+
+<div class="container-fluid text-center">
+    <h1>KOT: Keep On Track</h1>
+    <a class="m-2" th:href="@{/issue/new}">New Issue</a>
+    <a class="m-2" th:href="@{/issue/list}">Issue List</a>
+</div>
+
+<div class="container-fluid text-left">
+    <div><h2>New Issue:</h2></div>
+    <div>
+        <form th:action="@{/issue/save}" method="post" th:object="${issue}"
+            style="margin: 20px auto;">
+            <div class="form-group row">
+                <label class="col-sm-4 col-form-label">Author: </label>
+                <div class="col-sm-8">
+                    <input type="text" th:field="*{author}" class="form-control" required maxlength="50" />
+                </div>
+            </div>
+            <div class="form-group row">
+                <label class="col-sm-4 col-form-label">Description: </label>
+            </div>
+            <textarea rows="12" cols="40" type="" th:field="*{description}" class="form-control" required maxlength="200" />
+            <div class="text-center">
+                <button type="submit" class="btn btn-primary m-2">Save</button>
+                <button type="button" class="btn btn-secondary m-2" onclick="cancelForm()">Cancel</button>
+            </div>
+
+        </form>
+    </div>
+</div>
+
+<script type="text/javascript">
+    function cancelForm() {
+        window.location="[[@{/}]]"
+    }
+</script>
+
+</body>
+</html>
